Make IGameMetadata.getProperty returns Optional<String>
The use of optional is to force checks for non null value.
The reason for this is that getProperty() method will return null when a
key does not match. Optional forces to check for null values to minimize
potential null pointer exceptions and other common programming errors.
It also provides method so it becomes easier to do different actions
depending on if the value is null.

See: https://docs.oracle.com/javase/8/docs/api/java/util/Optional.html